Разбил админку на два приложения
