+++ django_admin/bot/apps.py
@@ -4,3 +4,4 @@ from django.apps import AppConfig
 class BotConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
     name = 'django_admin.bot'
+    verbose_name = "Внутренняя информация бота"
