+++ django_admin/bot/models.py
@@ -181,36 +181,3 @@ class Welcome_message(models.Model):
     )
     def __str__(self) -> str:
         return "Текст приветствия"
-
-    
-class ChatIDDirections(models.Model):
-    class Meta:
-        verbose_name = "ID чата"
-        verbose_name_plural = "ID чатов для вопросов по направлению подготовки"
-        app_label = "bot"
-        
-    chat_direction = models.OneToOneField(
-        "Directions",
-        verbose_name="Чат для направления подготовки:",
-        on_delete=CASCADE,
-    )
-    chat_id = models.BigIntegerField(
-        verbose_name="ID чата",
-        unique=True
-    )
-    
-    def __str__(self) -> str:
-        return f"{self.chat_direction}"
-    
-class ChatIDAdmission(models.Model):
-    class Meta:
-        verbose_name = "ID чата"
-        verbose_name_plural = "ID чата по поступлению"
-        app_label = "bot"
-        
-    chat_id = models.BigIntegerField(
-        verbose_name="ID чата",
-        unique=True
-    )
-    def __str__(self) -> str:
-        return "ChatID"
