+++ django_admin/django_admin/settings.py
@@ -39,7 +39,8 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'django_admin.bot'
+    'django_admin.bot.apps.BotConfig',
+    'django_admin.service.apps.ServiceConfig'
 ]
 
 MIDDLEWARE = [
