+++ django_admin/service/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+class ServiceConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'django_admin.service'
+    verbose_name = "Служебная информация"
